Move LoggingUtils
