+++ src/main/java/com/danielagapov/spawn/Services/BlockedUser/BlockedUserService.java
@@ -1,5 +1,6 @@
 package com.danielagapov.spawn.Services.BlockedUser;
 
+import com.danielagapov.spawn.DTOs.BlockedUser.BlockedUserDTO;
 import com.danielagapov.spawn.DTOs.FriendTag.FriendTagDTO;
 import com.danielagapov.spawn.Exceptions.Base.BaseNotFoundException;
 import com.danielagapov.spawn.Exceptions.Base.BaseSaveException;
@@ -10,14 +11,13 @@ import com.danielagapov.spawn.Models.User.User;
 import com.danielagapov.spawn.Repositories.User.IBlockedUserRepository;
 import com.danielagapov.spawn.Services.FriendTag.IFriendTagService;
 import com.danielagapov.spawn.Services.User.IUserService;
-import com.danielagapov.spawn.Utils.LoggingUtils;
+import com.danielagapov.spawn.Util.LoggingUtils;
 import org.springframework.cache.Cache;
 import org.springframework.cache.CacheManager;
-import org.springframework.dao.DataAccessException;
-import org.springframework.stereotype.Service;
-import com.danielagapov.spawn.DTOs.BlockedUser.BlockedUserDTO;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.Cacheable;
+import org.springframework.dao.DataAccessException;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.UUID;
@@ -42,11 +42,11 @@ public class BlockedUserService implements IBlockedUserService {
 
     @Override
     @CacheEvict(value = {
-        "blockedUsers", 
-        "blockedUserIds", 
-        "recommendedFriends", 
-        "otherProfiles", 
-        "friendsList"
+            "blockedUsers",
+            "blockedUserIds",
+            "recommendedFriends",
+            "otherProfiles",
+            "friendsList"
     }, key = "#blockerId")
     public void blockUser(UUID blockerId, UUID blockedId, String reason) {
         if (blockerId.equals(blockedId)) return;
@@ -54,13 +54,13 @@ public class BlockedUserService implements IBlockedUserService {
         try {
             User blocker = userService.getUserEntityById(blockerId);
             User blocked = userService.getUserEntityById(blockedId);
-            
-            logger.info("Attempting to block user: " + LoggingUtils.formatUserInfo(blocked) + 
-                         " by blocker: " + LoggingUtils.formatUserInfo(blocker));
-            
+
+            logger.info("Attempting to block user: " + LoggingUtils.formatUserInfo(blocked) +
+                    " by blocker: " + LoggingUtils.formatUserInfo(blocker));
+
             if (isBlocked(blockerId, blockedId)) {
-                logger.info("User " + LoggingUtils.formatUserInfo(blocked) + 
-                           " is already blocked by " + LoggingUtils.formatUserInfo(blocker));
+                logger.info("User " + LoggingUtils.formatUserInfo(blocked) +
+                        " is already blocked by " + LoggingUtils.formatUserInfo(blocker));
                 return;
             }
 
@@ -72,59 +72,59 @@ public class BlockedUserService implements IBlockedUserService {
             block.setReason(reason);
 
             repository.save(block);
-            
+
             // Also evict caches for the blocked user to ensure they don't see the blocker in recommendations
             evictBlockedUserCaches(blockedId);
-            
-            logger.info("Successfully blocked user: " + LoggingUtils.formatUserInfo(blocked) + 
-                         " by blocker: " + LoggingUtils.formatUserInfo(blocker));
+
+            logger.info("Successfully blocked user: " + LoggingUtils.formatUserInfo(blocked) +
+                    " by blocker: " + LoggingUtils.formatUserInfo(blocker));
         } catch (DataAccessException e) {
-            logger.error("Database error while blocking user " + LoggingUtils.formatUserIdInfo(blockedId) + 
-                         " by blocker " + LoggingUtils.formatUserIdInfo(blockerId) + ": " + e.getMessage());
+            logger.error("Database error while blocking user " + LoggingUtils.formatUserIdInfo(blockedId) +
+                    " by blocker " + LoggingUtils.formatUserIdInfo(blockerId) + ": " + e.getMessage());
             throw new BaseSaveException("Failed to block user: " + e.getMessage());
         } catch (BaseNotFoundException e) {
             logger.error("User not found while blocking: " + e.getMessage());
             throw e;
         } catch (Exception e) {
-            logger.error("Unexpected error while blocking user " + LoggingUtils.formatUserIdInfo(blockedId) + 
-                         " by blocker " + LoggingUtils.formatUserIdInfo(blockerId) + ": " + e.getMessage());
+            logger.error("Unexpected error while blocking user " + LoggingUtils.formatUserIdInfo(blockedId) +
+                    " by blocker " + LoggingUtils.formatUserIdInfo(blockerId) + ": " + e.getMessage());
             throw e;
         }
     }
 
     @Override
     @CacheEvict(value = {
-        "blockedUsers", 
-        "blockedUserIds", 
-        "recommendedFriends", 
-        "otherProfiles", 
-        "friendsList"
+            "blockedUsers",
+            "blockedUserIds",
+            "recommendedFriends",
+            "otherProfiles",
+            "friendsList"
     }, key = "#blockerId")
     public void unblockUser(UUID blockerId, UUID blockedId) {
         try {
             User blocker = userService.getUserEntityById(blockerId);
             User blocked = userService.getUserEntityById(blockedId);
-            
-            logger.info("Attempting to unblock user: " + LoggingUtils.formatUserInfo(blocked) + 
-                       " by blocker: " + LoggingUtils.formatUserInfo(blocker));
-                       
+
+            logger.info("Attempting to unblock user: " + LoggingUtils.formatUserInfo(blocked) +
+                    " by blocker: " + LoggingUtils.formatUserInfo(blocker));
+
             repository.findByBlocker_IdAndBlocked_Id(blockerId, blockedId)
                     .ifPresent(blockEntity -> {
                         repository.delete(blockEntity);
-                        
+
                         // Also evict caches for the previously blocked user
                         evictBlockedUserCaches(blockedId);
-                        
-                        logger.info("Successfully unblocked user: " + LoggingUtils.formatUserInfo(blocked) + 
-                                  " by blocker: " + LoggingUtils.formatUserInfo(blocker));
+
+                        logger.info("Successfully unblocked user: " + LoggingUtils.formatUserInfo(blocked) +
+                                " by blocker: " + LoggingUtils.formatUserInfo(blocker));
                     });
         } catch (DataAccessException e) {
-            logger.error("Database error while unblocking user " + LoggingUtils.formatUserIdInfo(blockedId) + 
-                         " by blocker " + LoggingUtils.formatUserIdInfo(blockerId) + ": " + e.getMessage());
+            logger.error("Database error while unblocking user " + LoggingUtils.formatUserIdInfo(blockedId) +
+                    " by blocker " + LoggingUtils.formatUserIdInfo(blockerId) + ": " + e.getMessage());
             throw e;
         } catch (Exception e) {
-            logger.error("Unexpected error while unblocking user " + LoggingUtils.formatUserIdInfo(blockedId) + 
-                         " by blocker " + LoggingUtils.formatUserIdInfo(blockerId) + ": " + e.getMessage());
+            logger.error("Unexpected error while unblocking user " + LoggingUtils.formatUserIdInfo(blockedId) +
+                    " by blocker " + LoggingUtils.formatUserIdInfo(blockerId) + ": " + e.getMessage());
             throw e;
         }
     }
@@ -133,16 +133,16 @@ public class BlockedUserService implements IBlockedUserService {
     @Cacheable(value = "isBlocked", key = "#blockerId.toString() + ':' + #blockedId.toString()")
     public boolean isBlocked(UUID blockerId, UUID blockedId) {
         try {
-            logger.info("Checking if user " + LoggingUtils.formatUserIdInfo(blockedId) + 
-                       " is blocked by " + LoggingUtils.formatUserIdInfo(blockerId));
+            logger.info("Checking if user " + LoggingUtils.formatUserIdInfo(blockedId) +
+                    " is blocked by " + LoggingUtils.formatUserIdInfo(blockerId));
             return repository.existsByBlocker_IdAndBlocked_Id(blockerId, blockedId);
         } catch (DataAccessException e) {
-            logger.error("Database error while checking if user " + LoggingUtils.formatUserIdInfo(blockedId) + 
-                         " is blocked by " + LoggingUtils.formatUserIdInfo(blockerId) + ": " + e.getMessage());
+            logger.error("Database error while checking if user " + LoggingUtils.formatUserIdInfo(blockedId) +
+                    " is blocked by " + LoggingUtils.formatUserIdInfo(blockerId) + ": " + e.getMessage());
             throw e;
         } catch (Exception e) {
-            logger.error("Unexpected error while checking if user " + LoggingUtils.formatUserIdInfo(blockedId) + 
-                         " is blocked by " + LoggingUtils.formatUserIdInfo(blockerId) + ": " + e.getMessage());
+            logger.error("Unexpected error while checking if user " + LoggingUtils.formatUserIdInfo(blockedId) +
+                    " is blocked by " + LoggingUtils.formatUserIdInfo(blockerId) + ": " + e.getMessage());
             throw e;
         }
     }
@@ -153,17 +153,17 @@ public class BlockedUserService implements IBlockedUserService {
         try {
             User blocker = userService.getUserEntityById(blockerId);
             logger.info("Getting blocked users for blocker: " + LoggingUtils.formatUserInfo(blocker));
-            
+
             List<BlockedUserDTO> blockedUsers = repository.findAllByBlocker_Id(blockerId).stream()
                     .map(BlockedUserMapper::toDTO)
                     .collect(Collectors.toList());
-                    
-            logger.info("Found " + blockedUsers.size() + " blocked users for blocker: " + 
-                       LoggingUtils.formatUserInfo(blocker));
+
+            logger.info("Found " + blockedUsers.size() + " blocked users for blocker: " +
+                    LoggingUtils.formatUserInfo(blocker));
             return blockedUsers;
         } catch (Exception e) {
-            logger.error("Error retrieving blocked users for blocker " + LoggingUtils.formatUserIdInfo(blockerId) + 
-                        ": " + e.getMessage());
+            logger.error("Error retrieving blocked users for blocker " + LoggingUtils.formatUserIdInfo(blockerId) +
+                    ": " + e.getMessage());
             throw e;
         }
     }
@@ -174,18 +174,18 @@ public class BlockedUserService implements IBlockedUserService {
         try {
             User blocker = userService.getUserEntityById(blockerId);
             logger.info("Getting blocked user IDs for blocker: " + LoggingUtils.formatUserInfo(blocker));
-            
+
             List<UUID> blockedUserIds = repository.findAllByBlocker_Id(blockerId).stream()
                     .map(BlockedUser::getBlocked)
                     .map(User::getId)
                     .collect(Collectors.toList());
-                    
-            logger.info("Found " + blockedUserIds.size() + " blocked user IDs for blocker: " + 
-                       LoggingUtils.formatUserInfo(blocker));
+
+            logger.info("Found " + blockedUserIds.size() + " blocked user IDs for blocker: " +
+                    LoggingUtils.formatUserInfo(blocker));
             return blockedUserIds;
         } catch (Exception e) {
-            logger.error("Error retrieving blocked user IDs for blocker " + LoggingUtils.formatUserIdInfo(blockerId) + 
-                        ": " + e.getMessage());
+            logger.error("Error retrieving blocked user IDs for blocker " + LoggingUtils.formatUserIdInfo(blockerId) +
+                    ": " + e.getMessage());
             throw e;
         }
     }
@@ -195,10 +195,10 @@ public class BlockedUserService implements IBlockedUserService {
         try {
             User userA = userService.getUserEntityById(userAId);
             User userB = userService.getUserEntityById(userBId);
-            
-            logger.info("Removing friendship between users: " + LoggingUtils.formatUserInfo(userA) + 
-                       " and " + LoggingUtils.formatUserInfo(userB));
-                       
+
+            logger.info("Removing friendship between users: " + LoggingUtils.formatUserInfo(userA) +
+                    " and " + LoggingUtils.formatUserInfo(userB));
+
             List<FriendTagDTO> userATags = friendTagService.getFriendTagsByOwnerId(userAId);
             for (FriendTagDTO tag : userATags) {
                 if (tag.getFriendUserIds().contains(userBId)) {
@@ -212,12 +212,12 @@ public class BlockedUserService implements IBlockedUserService {
                     friendTagService.removeUserFromFriendTag(tag.getId(), userAId);
                 }
             }
-            
-            logger.info("Successfully removed friendship between users: " + LoggingUtils.formatUserInfo(userA) + 
-                       " and " + LoggingUtils.formatUserInfo(userB));
+
+            logger.info("Successfully removed friendship between users: " + LoggingUtils.formatUserInfo(userA) +
+                    " and " + LoggingUtils.formatUserInfo(userB));
         } catch (Exception e) {
-            logger.error("Error removing friendship between users " + LoggingUtils.formatUserIdInfo(userAId) + 
-                        " and " + LoggingUtils.formatUserIdInfo(userBId) + ": " + e.getMessage());
+            logger.error("Error removing friendship between users " + LoggingUtils.formatUserIdInfo(userAId) +
+                    " and " + LoggingUtils.formatUserIdInfo(userBId) + ": " + e.getMessage());
             throw e;
         }
     }
