+++ src/main/java/com/danielagapov/spawn/Services/FriendRequest/FriendRequestService.java
@@ -16,11 +16,9 @@ import com.danielagapov.spawn.Models.User.User;
 import com.danielagapov.spawn.Repositories.IFriendRequestsRepository;
 import com.danielagapov.spawn.Services.BlockedUser.IBlockedUserService;
 import com.danielagapov.spawn.Services.User.IUserService;
-import com.danielagapov.spawn.Utils.LoggingUtils;
+import com.danielagapov.spawn.Util.LoggingUtils;
 import org.springframework.cache.CacheManager;
-import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.Cacheable;
-import org.springframework.cache.annotation.Caching;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
@@ -61,9 +59,9 @@ public class FriendRequestService implements IFriendRequestService {
 
             User sender = userService.getUserEntityById(senderId);
             User receiver = userService.getUserEntityById(receiverId);
-            
-            logger.info("Creating friend request from sender: " + LoggingUtils.formatUserInfo(sender) + 
-                        " to receiver: " + LoggingUtils.formatUserInfo(receiver));
+
+            logger.info("Creating friend request from sender: " + LoggingUtils.formatUserInfo(sender) +
+                    " to receiver: " + LoggingUtils.formatUserInfo(receiver));
 
             // Map the DTO to entity
             FriendRequest friendRequest = FriendRequestMapper.toEntity(friendRequestDTO, sender, receiver);
@@ -78,12 +76,12 @@ public class FriendRequestService implements IFriendRequestService {
             // Return the saved friend request DTO with additional details (friends and friend tags)
             return FriendRequestMapper.toDTO(friendRequest);
         } catch (DataAccessException e) {
-            logger.error("Failed to save friend request from user " + LoggingUtils.formatUserIdInfo(friendRequestDTO.getSenderUserId()) + 
-                         " to user " + LoggingUtils.formatUserIdInfo(friendRequestDTO.getReceiverUserId()) + ": " + e.getMessage());
+            logger.error("Failed to save friend request from user " + LoggingUtils.formatUserIdInfo(friendRequestDTO.getSenderUserId()) +
+                    " to user " + LoggingUtils.formatUserIdInfo(friendRequestDTO.getReceiverUserId()) + ": " + e.getMessage());
             throw new BaseSaveException("Failed to save friend request: " + e.getMessage());
         } catch (Exception e) {
-            logger.error("Error creating friend request from user " + LoggingUtils.formatUserIdInfo(friendRequestDTO.getSenderUserId()) + 
-                         " to user " + LoggingUtils.formatUserIdInfo(friendRequestDTO.getReceiverUserId()) + ": " + e.getMessage());
+            logger.error("Error creating friend request from user " + LoggingUtils.formatUserIdInfo(friendRequestDTO.getSenderUserId()) +
+                    " to user " + LoggingUtils.formatUserIdInfo(friendRequestDTO.getReceiverUserId()) + ": " + e.getMessage());
             throw e;
         }
     }
@@ -94,7 +92,7 @@ public class FriendRequestService implements IFriendRequestService {
         try {
             User user = userService.getUserEntityById(id);
             logger.info("Getting incoming fetch friend requests for user: " + LoggingUtils.formatUserInfo(user));
-            
+
             List<FriendRequest> friendRequests = getIncomingFriendRequestsByUserId(id);
             List<UUID> blockedUserIds = blockedUserService.getBlockedUserIds(id);
 
@@ -103,7 +101,7 @@ public class FriendRequestService implements IFriendRequestService {
                     .map(fr -> FetchFriendRequestMapper.toDTO(fr,
                             userService.getMutualFriendCount(id, fr.getSender().getId())))
                     .toList();
-                    
+
             logger.info("Found " + result.size() + " incoming fetch friend requests for user: " + LoggingUtils.formatUserInfo(user));
             return result;
         } catch (Exception e) {
@@ -117,18 +115,18 @@ public class FriendRequestService implements IFriendRequestService {
         try {
             List<FriendRequest> friendRequests = getIncomingFriendRequestsByUserId(id);
             List<CreateFriendRequestDTO> result = FriendRequestMapper.toDTOList(friendRequests);
-            
+
             return result;
         } catch (Exception e) {
             logger.error("Error retrieving incoming create friend requests for user: " + LoggingUtils.formatUserIdInfo(id) + ": " + e.getMessage());
             throw e;
         }
     }
-    
+
     public List<FriendRequest> getIncomingFriendRequestsByUserId(UUID id) {
         try {
             List<FriendRequest> requests = repository.findByReceiverId(id);
-            
+
             return requests;
         } catch (DataAccessException e) {
             logger.error("Database access error while retrieving incoming friend requests for user: " + LoggingUtils.formatUserIdInfo(id));
@@ -145,10 +143,10 @@ public class FriendRequestService implements IFriendRequestService {
             FriendRequest fr = repository.findById(id).orElseThrow(() -> new BaseNotFoundException(EntityType.FriendRequest, id));
             User sender = fr.getSender();
             User receiver = fr.getReceiver();
-            
-            logger.info("Accepting friend request with ID: " + id + " from sender: " + 
-                        LoggingUtils.formatUserInfo(sender) + " to receiver: " + LoggingUtils.formatUserInfo(receiver));
-                        
+
+            logger.info("Accepting friend request with ID: " + id + " from sender: " +
+                    LoggingUtils.formatUserInfo(sender) + " to receiver: " + LoggingUtils.formatUserInfo(receiver));
+
             userService.saveFriendToUser(sender.getId(), receiver.getId());
 
             // Publish friend request accepted notification event
@@ -184,12 +182,12 @@ public class FriendRequestService implements IFriendRequestService {
             if (fr != null) {
                 User sender = fr.getSender();
                 User receiver = fr.getReceiver();
-                logger.info("Deleting friend request with ID: " + id + " from sender: " + 
-                            LoggingUtils.formatUserInfo(sender) + " to receiver: " + LoggingUtils.formatUserInfo(receiver));
+                logger.info("Deleting friend request with ID: " + id + " from sender: " +
+                        LoggingUtils.formatUserInfo(sender) + " to receiver: " + LoggingUtils.formatUserInfo(receiver));
             } else {
                 logger.info("Deleting friend request with ID: " + id + " (request details not available)");
             }
-            
+
             repository.deleteById(id);
             logger.info("Friend request deleted successfully");
         } catch (DataAccessException e) {
@@ -203,19 +201,19 @@ public class FriendRequestService implements IFriendRequestService {
         try {
             User sender = userService.getUserEntityById(senderId);
             User receiver = userService.getUserEntityById(receiverId);
-            
-            logger.info("Deleting friend requests between sender: " + LoggingUtils.formatUserInfo(sender) + 
-                        " and receiver: " + LoggingUtils.formatUserInfo(receiver));
-                        
+
+            logger.info("Deleting friend requests between sender: " + LoggingUtils.formatUserInfo(sender) +
+                    " and receiver: " + LoggingUtils.formatUserInfo(receiver));
+
             List<FriendRequest> requests = repository.findBySenderIdAndReceiverId(senderId, receiverId);
             for (FriendRequest fr : requests) {
                 repository.delete(fr);
             }
-            
+
             logger.info("Deleted " + requests.size() + " friend requests between users");
         } catch (Exception e) {
-            logger.error("Error deleting friend request from user " + LoggingUtils.formatUserIdInfo(senderId) + 
-                         " to user " + LoggingUtils.formatUserIdInfo(receiverId) + ": " + e.getMessage());
+            logger.error("Error deleting friend request from user " + LoggingUtils.formatUserIdInfo(senderId) +
+                    " to user " + LoggingUtils.formatUserIdInfo(receiverId) + ": " + e.getMessage());
             throw e;
         }
     }
@@ -224,7 +222,7 @@ public class FriendRequestService implements IFriendRequestService {
     public List<FetchFriendRequestDTO> convertFriendRequestsToFetchFriendRequests(List<CreateFriendRequestDTO> friendRequests) {
         try {
             logger.info("Converting " + friendRequests.size() + " friend requests to fetch friend requests");
-            
+
             List<FetchFriendRequestDTO> fullFriendRequests = new ArrayList<>();
             for (CreateFriendRequestDTO friendRequest : friendRequests) {
                 UUID senderId = friendRequest.getSenderUserId();
@@ -238,7 +236,7 @@ public class FriendRequestService implements IFriendRequestService {
                         mutualFriendCount
                 ));
             }
-            
+
             logger.info("Converted " + fullFriendRequests.size() + " friend requests successfully");
             return fullFriendRequests;
         } catch (Exception e) {
@@ -251,10 +249,10 @@ public class FriendRequestService implements IFriendRequestService {
     public List<CreateFriendRequestDTO> getSentFriendRequestsByUserId(UUID userId) {
         try {
             User user = userService.getUserEntityById(userId);
-            
+
             List<FriendRequest> friendRequests = repository.findBySenderId(userId);
             List<CreateFriendRequestDTO> dtos = FriendRequestMapper.toDTOList(friendRequests);
-            
+
             return dtos;
         } catch (Exception e) {
             logger.error("Error retrieving sent friend requests for user: " + LoggingUtils.formatUserIdInfo(userId) + ": " + e.getMessage());
@@ -266,17 +264,17 @@ public class FriendRequestService implements IFriendRequestService {
     public Instant getLatestFriendRequestTimestamp(UUID userId) {
         try {
             logger.info("Getting latest friend request timestamp for user: " + LoggingUtils.formatUserIdInfo(userId));
-            
+
             // First check incoming friend requests
             Instant latestIncoming = repository.findTopByReceiverIdOrderByCreatedAtDesc(userId)
                     .map(FriendRequest::getCreatedAt)
                     .orElse(null);
-                    
+
             // Then check outgoing friend requests
             Instant latestOutgoing = repository.findTopBySenderIdOrderByCreatedAtDesc(userId)
                     .map(FriendRequest::getCreatedAt)
                     .orElse(null);
-            
+
             // Return the most recent timestamp, or null if no requests exist
             if (latestIncoming == null && latestOutgoing == null) {
                 return null;
