+++ src/main/java/com/danielagapov/spawn/Util/LoggingUtils.java
@@ -1,4 +1,4 @@
-package com.danielagapov.spawn.Utils;
+package com.danielagapov.spawn.Util;
 
 import com.danielagapov.spawn.Models.User.User;
 
@@ -8,10 +8,10 @@ import java.util.UUID;
  * Utility methods for consistent logging across the application
  */
 public class LoggingUtils {
-    
+
     /**
      * Format user information for logging, including ID, first name, last name, and username
-     * 
+     *
      * @param user The user entity
      * @return Formatted string with user details
      */
@@ -19,14 +19,14 @@ public class LoggingUtils {
         if (user == null) {
             return "null user";
         }
-        return String.format("%s with names: %s %s and username: %s", 
-            user.getId(), user.getFirstName(), user.getLastName(), user.getUsername());
+        return String.format("%s with names: %s %s and username: %s",
+                user.getId(), user.getFirstName(), user.getLastName(), user.getUsername());
     }
-    
+
     /**
      * Format user ID for logging when only the ID is available
      * Include a note that the full user info is not available
-     * 
+     *
      * @param userId The user ID
      * @return Formatted string with user ID
      */
